requirements.txt
streamlit
tensorflow
tf-keras
numpy
librosa
opencv-python
Pillow
matplotlib
scikit-image
lime
shap
pandas
plotly